Rename Image to mainImage
