+++ src/components/mainImage.js
