+++ src/pages/index.js
@@ -2,9 +2,10 @@ import React from "react"
 import { useStaticQuery, graphql } from "gatsby"
 
 import Layout from "../components/layout"
-import Image from "../components/image"
+import Image from "../components/mainImage"
 import SEO from "../components/seo"
 import Contact from "../components/contact"
+import Header from "../components/header"
 import { ThemeToggler } from 'gatsby-plugin-dark-mode'
 
 const IndexPage = () => {
@@ -17,9 +18,11 @@ const IndexPage = () => {
       }
     }
   `)
+  const title = data.site.siteMetadata.title
   return (
     <Layout>
       <SEO title="Home" />
+      <Header siteTitle={title}/>
       <div style={{ maxWidth: `300px`, marginBottom: `1.45rem`, marginTop: `2rem` }}>
         <Image />
         <ThemeToggler>
@@ -35,7 +38,7 @@ const IndexPage = () => {
           )}
         </ThemeToggler>
       </div>
-      <h1>{data.site.siteMetadata.title}</h1>
+      <h1>{title}</h1>
       <Contact />
     </Layout>
   )
